⚡️ feat: TSify Sapper's input files (client.js, server.js, service-worker.js → .ts) fix linting banning // @ts-ignore, remove unnecessary usage of it, and explain every one that remains
