+++ .eslintrc.json
@@ -15,6 +15,14 @@
 	},
 	"rules": {
 		"class-methods-use-this": "off",
+		"import/extensions": [
+			"error",
+			"ignorePackages",
+			{
+				"js": "never",
+				"ts": "never"
+			}
+		],
 		"import/prefer-default-export": "off",
 		"import/no-extraneous-dependencies": "off",
 		"indent": [
@@ -26,6 +34,13 @@
 			"error",
 			{
 				"allowIndentationTabs": true
+			}
+		],
+		"no-unused-vars": [
+			"error", 
+			{ 
+				"argsIgnorePattern": "^_",
+				"varsIgnorePattern": "^_"
 			}
 		],
 		"quotes": [
@@ -40,11 +55,15 @@
 			],
 			"parser": "@typescript-eslint/parser",
 			"extends": [
+				"plugin:import/typescript",
 				"plugin:@typescript-eslint/recommended"
 			],
 			"plugins": [
 				"@typescript-eslint"
-			]
+			],
+			"rules": {
+				"@typescript-eslint/ban-ts-comment": "off"
+			}
 		},
 		{
 			"files": [
