+++ package.json
@@ -16,7 +16,7 @@
 		"url": "https://github.com/babichjacob/sapper-typescript-graphql-template.git"
 	},
 	"license": "MIT",
-	"version": "2020.05.27",
+	"version": "2020.05.28",
 	"scripts": {
 		"eslint": "eslint",
 		"eslint:fix": "eslint --fix ./*.js ./src/*.js ./src/components/*.svelte ./src/graphql/*.ts ./src/routes/*.svelte ./src/routes/*.ts",
@@ -30,7 +30,7 @@
 		"compression": "^1.7.4",
 		"express": "^4.17.1",
 		"node-fetch": "^2.6.0",
-		"sirv": "^0.4.6"
+		"sirv": "^1.0.0-next.5"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.10.0",
@@ -44,6 +44,7 @@
 		"@rollup/plugin-node-resolve": "^8.0.0",
 		"@rollup/plugin-replace": "^2.2.0",
 		"@rollup/plugin-typescript": "^4.1.2",
+		"@types/compression": "^1.7.0",
 		"@types/express": "^4.17.6",
 		"@types/node-fetch": "^2.5.7",
 		"@typescript-eslint/eslint-plugin": "^3.0.1",
