--- src/client.js
@@ -1,5 +1,0 @@
-import * as sapper from "@sapper/app";
-
-sapper.start({
-	target: document.querySelector("#sapper"),
-});
