+++ src/client.ts
@@ -1,0 +1,6 @@
+// @ts-ignore -- generated package
+import * as sapper from "@sapper/app"; // eslint-disable-line import/no-unresolved
+
+sapper.start({
+	target: document.querySelector("#sapper"),
+});
