+++ src/routes/_error.svelte
@@ -1,8 +1,8 @@
 <script lang="typescript">
 	export let status: string;
 	export let error: Error;
-	
-	// @ts-ignore
+
+	// @ts-ignore -- creates a warning when the left side of the === is replaced with "production" by rollup-plugin-replace in `rollup.config.js`
 	const dev = process.env.NODE_ENV === "development";
 </script>
